+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#792](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/792))
 - Add support for generic OTEL_PYTHON_EXCLUDED_URLS variable
   ([#790](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/790))
+ - `opentelemetry-instrumentation-kafka-python` added kafka-python module instrumentation.
+  ([#TODO](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/TODO))
 
 ### Fixed